Make the native share menu primary and the custom one fallback
Most devices support a native share menu. However, that menu might not
be present at times and it is not available if the network is not HTTPS,
for cases like that the fallback custom menu is set that will use the
modal in order to let the user copy the link.